Move GetEntityByIDHandler and DeleteEntityByIDHandler to entity_handlers.go file
